--- a/pkg/sql/opt/memo/expr_view.go
+++ b/pkg/sql/opt/memo/expr_view.go
@@ -107,6 +107,11 @@ func (ev ExprView) Operator() opt.Operator {
 	return ev.op
 }
 
+// Memo returns the Memo structure over which the ExprView operates.
+func (ev ExprView) Memo() *Memo {
+	return ev.mem
+}
+
 // Logical returns the set of logical properties that this expression provides.
 func (ev ExprView) Logical() *props.Logical {
 	return &ev.mem.group(ev.group).logical
@@ -199,8 +204,7 @@ func (ev ExprView) Replace(evalCtx *tree.EvalContext, replace ReplaceChildFunc) 
 	}
 	existingExpr := ev.mem.NormExpr(ev.group)
 	newExpr := existingExpr.Replace(ev.mem, replace)
-	newGroup := ev.mem.MemoizeNormExpr(evalCtx, newExpr)
-	return MakeNormExprView(ev.mem, newGroup)
+	return ev.mem.MemoizeNormExpr(evalCtx, newExpr)
 }
 
 func (ev ExprView) childGroup(nth int) *group {