… pkg/sql/opt/bench/bench_test.go
… pkg/sql/opt/bench/stub_factory.go
… pkg/sql/opt/idxconstraint/index_constraints_test.go
… pkg/sql/opt/memo/list_storage_test.go
… pkg/sql/opt/memo/memo_test.go
… pkg/sql/opt/metadata_test.go
… pkg/sql/opt/norm/factory.go
… pkg/sql/opt/norm/testdata/rules/prune_cols
… pkg/sql/opt/norm/testdata/rules/select
… pkg/sql/opt/optbuilder/builder.go
… pkg/sql/opt/optbuilder/orderby.go
… pkg/sql/opt/optbuilder/scalar.go
… pkg/sql/opt/optbuilder/select.go
… pkg/sql/opt/optbuilder/util.go
… pkg/sql/opt/optgen/cmd/optgen/factory_gen.go
… pkg/sql/opt/optgen/cmd/optgen/testdata/factory
… pkg/sql/opt/props/logical.go
… pkg/sql/opt/testutils/testcat/create_table.go
… pkg/sql/opt/testutils/testcat/create_view.go
… pkg/sql/opt/testutils/testcat/test_catalog.go
… pkg/sql/opt/xform/testdata/external/hibernate
… pkg/sql/opt/xform/testdata/external/nova
… pkg/sql/opt_catalog.go
… pkg/sql/opt_index_selection_test.go
… pkg/sql/pgwire/pgwire_test.go
… pkg/sql/plan.go
… pkg/sql/prepared_stmt.go
… pkg/sql/sessiondata/search_path.go
… pkg/sql/statement.go
… pkg/sql/table.go