Store the cost-based optimizer's Memo struct in the PreparedStatement,
and then use that to speed up the planning process on each execution,
since the `optbuild` and `normalize` phases can be either skipped or
significantly reduced. If there are no placeholders, then the `explore`
phase can be skipped as well.

This PR significantly improves execution times when plans have been
prepared. The "old time" is the planning time when only the AST is
cached. The "new time" is the planning time when both the AST and
Memo data structures are cached.
```
name                               old time/op  new time/op  delta
Phases/kv-read/ExecBuild           24.6µs ± 2%  16.2µs ± 1%  -34.38%  (p=0.008 n=5+5)
Phases/kv-read-const/ExecBuild     24.9µs ± 1%   0.7µs ± 2%  -97.35%  (p=0.008 n=5+5)
Phases/tpcc-new-order/ExecBuild    91.8µs ± 3%  56.5µs ± 1%  -38.53%  (p=0.008 n=5+5)
Phases/tpcc-delivery/ExecBuild     67.6µs ± 1%  41.7µs ± 2%  -38.39%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/ExecBuild   357µs ± 4%   182µs ± 1%  -48.87%  (p=0.008 n=5+5)
Exec/kv-read/ExecOpt                331µs ± 2%   187µs ± 2%  -43.67%  (p=0.008 n=5+5)
Exec/kv-read-const/ExecOpt          235µs ± 1%   160µs ± 1%  -31.83%  (p=0.008 n=5+5)
Exec/tpcc-new-order/ExecOpt         499µs ± 1%   275µs ± 1%  -44.77%  (p=0.008 n=5+5)
Exec/tpcc-delivery/ExecOpt          430µs ± 3%   243µs ± 1%  -43.49%  (p=0.008 n=5+5)
Exec/tpcc-stock-level/ExecOpt      1.03ms ± 1%  0.58ms ± 1%  -43.27%  (p=0.008 n=5+5)
```
Notice that the `kv-read-const` can skip almost all of planning, since it
contains no placeholders. Consequently, it takes only 700ns to plan.

With this change, end-to-end running times (on empty tables) for the
cost-based optimizer are typically lower for prepared queries than the
heuristic optimizer:
```
name                        old time/op  new time/op  delta
Exec/kv-read/Exec            196µs ± 3%   187µs ± 2%   -4.78%  (p=0.008 n=5+5)
Exec/kv-read-const/Exec      198µs ± 1%   160µs ± 1%  -18.83%  (p=0.008 n=5+5)
Exec/tpcc-new-order/Exec     287µs ± 3%   275µs ± 1%   -4.04%  (p=0.008 n=5+5)
Exec/tpcc-delivery/Exec      245µs ± 2%   243µs ± 1%     ~     (p=0.421 n=5+5)
Exec/tpcc-stock-level/Exec   552µs ± 1%   583µs ± 1%   +5.54%  (p=0.008 n=5+5)
```
The `tpcc-stock-level` is a bit higher, since the cost-based optimizer is
currently missing constant-folding.